changed test splash to production splash
