+++ frontend/app/App.js
@@ -9,8 +9,6 @@ import Confirmation from './screens/Confirmation.js';
 import HeaderBanner from './screens/HeaderBanner.js';
 import Signup from './screens/Signup.js';
 import Splash from './screens/Splash.js';
-import testSplash from './screens/testSplash.js';
-
 
 import NewCardPage from './stripe_page.js';
 
@@ -22,7 +20,7 @@ const AppNavigator = StackNavigator(
 
   {
     Index: {
-      screen: testSplash
+      screen: Splash
     },
     NewCardPage: {
       screen: NewCardPage
