+++ frontend/app/screens/Splash.js
@@ -1,37 +1,69 @@
-import React from 'react';
+
+import React from 'react'
 import {
   Text,
-  View
+  View,
+  Animated,
+  Image,
+  Easing,
+  StyleSheet
 } from 'react-native';
+
 import {
   AccessToken
 } from 'react-native-fbsdk';
 
-// import * as firebase from 'firebase';
 
 class Splash extends React.Component {
-  componentDidMount() {
+  constructor () {
+    super()
+    this.springValue = new Animated.Value(0.3)
+  }
 
+  componentDidMount() {
+    this.spring()
     AccessToken.getCurrentAccessToken().then(token => {
       if (token) {
         // firebase.database().ref(`/users/${token.userID}`)
         //         .on('value', (snap) => this.props.receiveCurrentUser(snap.val()));
-        this.props.navigation.navigate('SimpleApp');
+        setTimeout(()=>  this.props.navigation.navigate('SimpleApp'), 2000)
       } else {
-        this.props.navigation.navigate('Signup');
+        setTimeout(()=>  this.props.navigation.navigate('Signup'), 2000)
       }
     });
   }
 
-  render() {
-    return (
-      <View>
-        <Text></Text>
-      </View>
-    );
+
+  spring() {
+    this.springValue.setValue(0.3)
+    Animated.spring(
+      this.springValue,
+      {
+        toValue: 1,
+        friction: 1,
+      }
+    ).start()
   }
 
+  render () {
+
+    return (
+      <View style={styles.container}>
+        <Animated.Image
+          style={{ width: 227, height: 200, transform: [{scale: this.springValue}] }}
+          source={{uri: 'https://s3.amazonaws.com/media-p.slid.es/uploads/alexanderfarennikov/images/1198519/reactjs.png'}}/>
+      </View>
+    )
+  }
 }
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center'
+  }
+})
 
-export default Splash;
+
+export default Splash
